requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.54.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.147.0
sqlalchemy==2.0.35
pyyaml==6.0.2
# CMS
jinja2==3.1.4
python-multipart==0.0.12
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
